agents/support/support-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "support-agent"
version = "1.0.0"
description = "Support agent for automated customer support"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "anthropic>=0.25.0",
    "qdrant-client>=1.9.0",
    "fastapi>=0.110.0",
    "psycopg2-binary",
    "redis[hiredis]>=4.6.0",
    "resend>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]